update Manifest to account for the fact that position indexes can exceed the max int value

+++ concourse-server/src/main/java/com/cinchapi/concourse/server/io/NullByteSink.java
@@ -28,7 +28,7 @@ final class NullByteSink implements ByteSink {
     /**
      * The tracked "position".
      */
-    private int position;
+    private long position;
 
     @Override
     public long position() {
